Limitation of receiving data from the server.
A check has been added to ensure that data received from the server must not be less than the already loaded data. If the downloaded data is less than the data on the server, the “show more” button will be displayed, but if it is greater than or equal to , the button will be hidden.