Deploying motion properties.
A separate method has been created that returns general framer motion properties.
Fixed animation delay in the custom property.
A Boolean parameter has been added to the method that returns the framer motion properties, which answers when the animation will play, when it will be in the viewport or not.